Introduced proper script file parsing
Testgearctl now parses the provided test script and it does so by
dividing the contents of the test script into chunks. Each chunk is
considered a test case. The start of each chunk is marked with the
following lua comment:

"-- tg" (without quotation marks)

In the future, this will make it possible to traverse the test cases and
print a pass/fail result for each test case.

This commit also introduces the use of luaprompt v0.3.1 which provides a
more colorful and easier to use lua prompt in interactive mode.