Add support for printing help strings using '?'
This patch adds support for printing descriptive help strings when a '?'
is appended.

If questioning a table it simply looks up and prints the value of the
key string named _self_help of the same table.

If questioning a non-table it will try to look up the name of the
corresponding key in the _help table of the parent if present.

Example:

dummy = { _help_self = "I'm a dummy", int0 = 42 }
dummy._help = { int0 = "Integer 0 description" }

$ dummy ?
I'm a dummy

$ dummy.int0 ?
Integer 0 description

$ dummy.int0
42